Include various missed dependencies
The reason for the prior bug 'cannot find %>%'  was that it was not included in the package.

The 'fix' of `library(tidyverse)` covered this off on a per session basis, but also then covered subsequent issues about locating `rowsie()` and `mutate()` which also hadn't been namespaced.

Runnning `usethis::use_pipe()` creates the required file in 'R/utils-pipe.R' and change to description.

The `R/badger.R` file was extended to include namespaced dplyr functions, and teh docs then extended with the recommended output from `sinew::makeOxy(create_badge())`